Added Netbeans project files
Here is the environment framework I've started. Most of the classes
necessary for the GUI portion of the project are here, and I will flesh
them out as we collectively decide on details of the project. Still
missing are classes for the rover itself and classes for coding the
inference engine and other KBS-oriented parts of the project.